Expanded capabilities for rio shapes.
Closes #150.

Switch CLI over to faster transforms.

Also add option for output precision.

Closes #163.
