+++ tests/test_warp.py
@@ -7,7 +7,8 @@ import affine
 import numpy
 
 import rasterio
-from rasterio.warp import reproject, RESAMPLING
+from rasterio.warp import reproject, RESAMPLING, transform_geom
+
 
 logging.basicConfig(stream=sys.stderr, level=logging.DEBUG)
 
@@ -175,3 +176,9 @@ def test_warp_from_to_file_multi(tmpdir):
                     num_threads=2)
     # subprocess.call(['open', tiffname])
 
+def test_transform_geom_wrap():
+    geom = {'type': 'Polygon', 'coordinates': (((798842.3090855901, 6569056.500655151), (756688.2826828464, 6412397.888771972), (755571.0617232556, 6408461.009397383), (677605.2284582685, 6425600.39266733), (677605.2284582683, 6425600.392667332), (670873.3791649605, 6427248.603432341), (664882.1106069803, 6407585.48425362), (663675.8662823177, 6403676.990080649), (485120.71963574126, 6449787.167760638), (485065.55660851026, 6449802.826920689), (485957.03982722526, 6452708.625101285), (487541.24541826674, 6457883.292107048), (531008.5797472061, 6605816.560367976), (530943.7197027118, 6605834.9333479265), (531888.5010308184, 6608940.750411527), (533299.5981959199, 6613962.642851984), (533403.6388841148, 6613933.172096095), (576345.6064638699, 6761983.708069147), (577649.6721159086, 6766698.137844516), (578600.3589008929, 6770143.99782289), (578679.4732294685, 6770121.638265098), (655836.640492081, 6749376.357102599), (659913.0791150068, 6764770.1314677475), (661105.8478791204, 6769515.168134831), (661929.4670843681, 6772800.8565198565), (661929.4670843673, 6772800.856519875), (661975.1582566603, 6772983.354777632), (662054.7979028501, 6772962.86384242), (841909.6014891531, 6731793.200435557), (840726.455490463, 6727039.8672589315), (798842.3090855901, 6569056.500655151)),)}
+    result = transform_geom(
+                'EPSG:3373', 'EPSG:4326', geom, antimeridian_cutting=True)
+    assert result['type'] == 'MultiPolygon'
+    assert len(result['coordinates']) == 2
